+++ lib/builder.dart
@@ -1,12 +1,8 @@
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
-import 'package:built_value_generator/built_value_generator.dart';
 
 import 'generator.dart';
 
-Builder builtRedux(BuilderOptions options) => new PartBuilder([
+Builder builtRedux(BuilderOptions _) => new SharedPartBuilder([
       new BuiltReduxGenerator(),
-      new BuiltValueGenerator(),
-    ],
-        generatedExtension: '.g.dart',
-        header: options.config['header'] as String);
+    ], 'built_redux');